requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudinary==1.20.0
colorama==0.4.3
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.9.0
django-environ==0.4.5
django-heroku==0.3.1
get==2019.4.13
gunicorn==20.0.4
idna==2.9
isort==4.3.21
langdetect==1.0.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
post==2019.4.13
psycopg2==2.7.7
psycopg2-binary==2.8.5
public==2019.4.13
pycparser==2.20
pylint==2.4.4
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
six==1.14.0
soupsieve==2.0
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
